requirements.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cmake==3.18.0
distlib==0.3.1
Django==3.0.6
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
idna==2.10
Markdown==3.3.3
mysqlclient @ file:///D:/DriveC/PycharmProjects/readingright3/mysqlclient-1.4.6-cp38-cp38-win32.whl
numpy==1.19.1
Pillow==7.2.0
pipenv==2020.11.15
PyJWT==2.0.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
virtualenv==20.0.25
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
